test: add test_utils crate and move test_valid_config_body to it
